--- mutis/correlation/__init__.py
@@ -1,5 +1,0 @@
-# Licensed under a 3-clause BSD style license - see LICENSE
-"""Correlation analysis."""
-
-from .core import *
-from .functions import *
