--- mutis/utils/__init__.py
@@ -1,11 +1,0 @@
-# Licensed under a 3-clause BSD style license - see LICENSE
-"""
-Utility functions and classes used throughout MUTIS.
-
-You have to import sub-modules of `mutis.utils` directly,
-the `mutis.utils` namespace is empty.
-
-Examples::
-
-    from mutis.utils.utils import memoize
-"""
